refactor: improve reusabillity
